requirements.txt
py-cord~=2.5.0
PyYAML~=6.0.1
python-dotenv~=1.0.1
requests~=2.31.0